have running extensions use the issure reporter instead
