+++ src/vs/workbench/contrib/extensions/browser/browserRuntimeExtensionsEditor.ts
@@ -7,6 +7,7 @@ import { Action } from 'vs/base/common/actions';
 import { IExtensionHostProfile } from 'vs/workbench/services/extensions/common/extensions';
 import { ExtensionIdentifier } from 'vs/platform/extensions/common/extensions';
 import { AbstractRuntimeExtensionsEditor, IRuntimeExtension } from 'vs/workbench/contrib/extensions/browser/abstractRuntimeExtensionsEditor';
+import { ReportExtensionIssueAction } from 'vs/workbench/contrib/extensions/common/reportExtensionIssueAction';
 
 export class RuntimeExtensionsEditor extends AbstractRuntimeExtensionsEditor {
 
@@ -23,6 +24,9 @@ export class RuntimeExtensionsEditor extends AbstractRuntimeExtensionsEditor {
 	}
 
 	protected _createReportExtensionIssueAction(element: IRuntimeExtension): Action | null {
+		if (element.marketplaceInfo) {
+			return this._instantiationService.createInstance(ReportExtensionIssueAction, element.description);
+		}
 		return null;
 	}
